+++ RuntimeSpawner/RuntimeMultiFontTextRendererSpawner.cs
@@ -37,13 +37,13 @@ namespace TextMeshDOTS.Authoring
 
             textBaseConfiguration = new TextBaseConfiguration
             {
-                fontSize = 12,
+                fontSize = (half)12,
                 color = Color.white,
                 fontStyles = FontStyles.Normal,
                 maxLineWidth = 10,
-                wordSpacing = 0,
-                lineSpacing = 0,
-                paragraphSpacing = 0,
+                wordSpacing = (half)0,
+                lineSpacing = (half)0,
+                paragraphSpacing = (half)0,
                 lineJustification = HorizontalAlignmentOptions.Left,
                 verticalAlignment = VerticalAlignmentOptions.TopBase,
                 isOrthographic = false,
